+++ R/walsNB.R
@@ -924,8 +924,8 @@ print.summary.walsNB <- function(x, digits = max(3, getOption("digits") - 3), ..
 #' is estimated via \code{\link[MASS]{glm.nb}} and \code{initThetaMASS} is ignored.
 #' If \code{FALSE}, then the initial fit is estimated by minimizing the
 #' log-likelihood using \code{\link[stats]{optim}}.
-#' @param restricted If TRUE, then initial fit in \code{\link[WALS]{fitNB2}} only
-#' considers the focus regressors. By default \code{FALSE}, then the unrestricted
+#' @param restricted If \code{TRUE}, then initial fit in \code{\link[WALS]{fitNB2}}
+#' only considers the focus regressors. By default \code{FALSE}, then the unrestricted
 #' model is estimated in \code{\link[WALS]{fitNB2}} (i.e. all regressors).
 #' @param eps Controls argument \code{eps} in \code{\link[WALS]{fitNB2}} for generating
 #' starting value for \code{logTheta} (\eqn{\log{\theta}}) via \code{\link[MASS]{theta.ml}}.